requirements.txt
# Core Dependencies
langchain-text-splitters==0.3.2
google-generativeai==0.8.3

# Vector Database
chromadb==0.4.22

# Document Processing
PyMuPDF>=1.24.0
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.2.0

# Text Processing
sentence-transformers==2.3.1

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
numpy==1.26.3

# Testing
pytest==7.4.3
httpx==0.26.0
